+++ src/shared/ui/textarea/RadixTextarea.module.scss
@@ -3,58 +3,67 @@
   width: 100%;
   background-color: var(--dark-500);
 
-  & .textarea {
-    width: 100%;
-    min-height: 84px;
-    padding: 6px 12px;
-    border-radius: 2px;
-    font-family: var(--font-family);
-    background-color: var(--dark-500);
+  & .text {
 
-    font-size: 16px;
-    line-height: 150%;
-    color: var(--light-900);
-    border: 1px solid var(--dark-100);
-    outline: none;
-    resize: none;
-    transition: all .3s;
+  }
 
-    &::placeholder {
-      color: var(--light-900);
-    }
+  & .textareaWrapper {
 
-    &:hover {
-      border: 1px solid var(--dark-100);
-      color: var(--light-900);
-    }
-
-    &:focus {
-      border: 2px solid var(--primary-700);
-      color: var(--light-100);
-    }
-
-    &:active {
-      border: 1px solid var(--light-100);
-      color: var(--light-100);
-    }
-
-    &:disabled {
-      border: 1px solid var(--dark-100);
-      color: var(--dark-100);
+    textarea {
+      width: 100%;
+      min-height: 84px;
+      padding: 6px 12px;
+      border-radius: 2px;
+      font-family: var(--font-family);
       background-color: var(--dark-500);
-      cursor: not-allowed;
-      opacity: 1;
-      -webkit-text-fill-color: var(--light-700); // Для Safari
+
+      font-size: 16px;
+      line-height: 150%;
+      color: var(--light-900);
+      border: 1px solid var(--dark-100);
+      outline: none;
+      resize: none;
+      transition: all .3s;
+
+      &::placeholder {
+        color: var(--light-900);
+      }
+
+      &:hover {
+        border: 1px solid var(--dark-100);
+        color: var(--light-900);
+      }
+
+      &:focus {
+        border: 2px solid var(--primary-700);
+        color: var(--light-100);
+      }
+
+      &:active {
+        border: 1px solid var(--light-100);
+        color: var(--light-100);
+      }
+
+      &:disabled {
+        border: 1px solid var(--dark-100);
+        color: var(--dark-100);
+        background-color: var(--dark-500);
+        cursor: not-allowed;
+        opacity: 1;
+      }
     }
+
+
   }
 
 }
 
 
-
 .error {
-  border: 1px solid var(--danger-500) !important;
-  color: var(--light-100) !important;
+  textarea {
+    border: 1px solid var(--danger-500) !important;
+    color: var(--light-100) !important;
+  }
 }
 
 .errorMessage {
